rename tests/ to mdtest/
mdtest feels like a good name for this package.
