+++ mdtest/path_test.go
@@ -1,4 +1,4 @@
-package tests
+package mdtest
 
 import (
 	"path/filepath"
@@ -12,6 +12,6 @@ func TestRepoAbsPathLocation(t *testing.T) {
 	t.Parallel()
 	dir, err := RepoAbsPath()
 	require.Equal(t, nil, err)
-	f := filepath.Join(dir, "tests", "path.go")
-	assert.FileExists(t, f, "%s not in expected repo path pkg/test/path.go", f)
+	f := filepath.Join(dir, "mdtest", "path.go")
+	assert.FileExists(t, f, "%s not in expected repo path pkg/mdtest/path.go", f)
 }
