+++ mdtest/zq_example_test.go
@@ -1,4 +1,4 @@
-package tests
+package mdtest
 
 import (
 	"testing"
